arrays/sum.go
package main

// Takes a slice of integers and returns sum of them.
func Sum(numbers []int) int {
	var sum int

	for _, number := range numbers {
		sum += number
	}

	return sum
}

// Takes number of slices and returns slice of sum of tails of each slice
// tails is all the numbers in a slice apart from head(the first one)
func SumAllTails(numbersToSum ...[]int) []int {
	var sums []int

	for _, numbers := range numbersToSum {
		if len(numbers) == 0 {
			sums = append(sums, 0)
		} else {
			tail := numbers[1:]
			sums = append(sums, Sum(tail))
		}
	}

	return sums
}
